refactor(generation): split baseplateDirectMesh into focused modules
Decompose the 1084-LOC baseplateDirectMesh.ts into 6 sibling modules.
The orchestrator now contains only `generateBaseplateDirect`, which
wires together pure-geometry helpers and add* emitters.

Extracted modules (all under 250 LOC):

- directMeshBuilder.ts (150) — `MeshBuilder` class, `faceNormal`,
  `tangentVectors`, plus `CORNER_SEGMENTS`, `CIRCLE_SEGMENTS`,
  `CANCEL_EPSILON` constants.
- directMeshShapes.ts (162) — `roundedRectPoints`,
  `roundedRectPointsSelective`, `circlePoints`, `pointInPolygon`.
  All 2D shape primitives shared by walls, faces, and magnets.
- directMeshFaces.ts (218) — `addPlateFace` (top/bottom slab face),
  `addCellCornerGussets` (the per-cell corner fillet complements that
  tile the inter-pocket lattice), `addRingFace` (padding ring), and
  `addSolidBottomFace`.
- directMeshWalls.ts (116) — `addPocketWalls` (tapered walls + optional
  floor cap) and `addOuterWalls`.
- directMeshMagnets.ts (90) — `addMagnetHoles` (cancel + cylinder + floor
  pattern, smooth-shaded shared rings).
- directMeshConnectors.ts (162) — `addConnectorNub` (male protrusion)
  and `addConnectorHole` (female cancel-pattern indentation).

Public surface unchanged: `generateBaseplateDirect` continues to be the
sole export, with the same signature consumers expect.

baseplateDirectMesh.ts: 1084 LOC -> 205 LOC. All 10,278 tests pass.
